Builds.

+++ LiquidEngine/Graphics/Renderer/GraphicsDescriptorHeaps.cpp
@@ -22,8 +22,8 @@ void GraphicsDescriptorHeaps::clean_up() {
 
 UINT GraphicsDescriptorHeaps::get_open_heap_index() {
 	auto find = std::ranges::find(occupation, false);
-	if (find != occupation.end()) return std::distance(occupation.begin(), find);
-	else return occupation.size();
+	if (find != occupation.end()) return (UINT)std::distance(occupation.begin(), find);
+	else return (UINT)occupation.size();
 }
 
 void GraphicsDescriptorHeaps::create_cbv(const ComPtr<ID3D12Device> &device, const ComPtr<ID3D12Resource> &upload_buffer, size_t obj_size, UINT frame_index, UINT heap_index) {
